+++ Peer/App/AppBuilder/HandlerWrapper.cs
@@ -3,18 +3,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Peer.Parsing.CommandLine;
-
-public interface IHandler<in TVerb>
-{
-    Task HandleAsync(TVerb opts, IServiceCollection collection, CancellationToken token = default);
-}
-
-public interface IHandler
-{
-    Task HandleAsync(object opts, IServiceCollection collection, CancellationToken token = default);
-    Type Type { get; }
-}
+namespace Peer.App.AppBuilder;
 
 public class HandlerWrapper<T> : IHandler
 {
@@ -35,4 +24,3 @@ public class HandlerWrapper<T> : IHandler
         return HandleAsync((T)opts, collection, token);
     }
 }
-
