+++ Peer/App/IVerb.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using Peer.Verbs;
 
-namespace Peer.Parsing.CommandLine;
+namespace Peer.App;
 
 public interface IVerb
 {
