+++ Peer/App/SubVerb.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using Peer.Verbs;
 
-namespace Peer.Parsing.CommandLine;
+namespace Peer.App;
 
 public class SubVerb<TVerb, TSubVerb> : Verb<TSubVerb>, ISubVerb<TVerb>
 {
